--- wct_app/README.md
@@ -1,6 +1,0 @@
-# wct-tista
-
-https://github.com/enriquemanuel/wct-tista/invitations
-
-https://circleci.com/gh/enriquemanuel/wct-tista/
-
