--- wct_app/requirements.txt
@@ -1,41 +1,0 @@
-astroid==2.2.5
-atomicwrites==1.3.0
-attrs==19.1.0
-certifi==2019.6.16
-chardet==3.0.4
-Click==7.0
-dparse==0.4.1
-Flask==1.0.2
-gunicorn==19.9.0
-idna==2.8
-importlib-metadata==0.19
-isort==4.3.21
-itsdangerous==1.1.0
-Jinja2==2.10.1
-joblib==0.13.2
-lazy-object-proxy==1.4.1
-MarkupSafe==1.1.1
-mccabe==0.6.1
-more-itertools==7.2.0
-packaging==19.1
-pluggy==0.12.0
-py==1.8.0
-pylint==2.3.1
-pyparsing==2.4.2
-pytest==5.0.1
-python-dateutil==2.8.0
-pytz==2019.2
-PyYAML==5.1.2
-requests==2.22.0
-safety==1.8.5
-six==1.12.0
-sklearn==0.0
-sqlparse==0.3.0
-typed-ast==1.4.0
-urllib3==1.25.3
-virtualenv==16.4.3
-wcwidth==0.1.7
-Werkzeug==0.15.5
-wrapt==1.11.2
-xlrd==1.2.0
-zipp==0.5.2
