+++ wct_app/templates/help.html
@@ -1,0 +1,21 @@
+<h1>A Guide to Solving Web Development Problems</h1>
+
+<p>This is mainly about how to solve technical problems that arise from using front or back end technologies to make web pages or web apps but some of these steps will be applicable to solving technical problems in general.</p>
+
+<p>Half the technical problems in web development are caused by something trivial like forgetting a semi-colon somewhere and are usually easily solvable if you've had your coffee but for all the problems past this level, you'll probably need to do some structured thinking.</p>
+
+<h2> The Web Development Troubleshooting Steps Summary</h2>
+
+<p>1. Define the problem that you're trying to solve. If you're not asking exactly the right question then you're not going to get exactly the right answer.</p>
+
+<p>2. Verify the problem exists. Make sure it's really a problem by replicating the error on more than one machine.</p>
+
+<p>3. Gather additional information about the problem. Use a javascript debugger, fiddler or any other common tool to get additional clues about the nature of the problem.</p>
+
+<p>4. Back-up your work before making even a single teeny tiny little change. Don't make things worse when trying to fix the problem. Keep a back-up so you can roll back if you have to.</p>
+
+<p>5. Review your past experiences versus the present situation. Have you encountered a similar problem in the past, if so then test to see if your past approach will fix the problem at hand.</p>
+
+<p>6. Search for people who've been in similar situations and how they solved the problem. You have all the information you need so start using a search engine to find possible solutions.</p>
+
+<p>7. Make one change at a time (and then test it). Proceed methodically and if you make multiple changes at the same time then you often can't tell what worked or didn't work. So proceed step by step.</p>
